feat: add admin panel modal
